move beaconStateReader
